+++ lib/generation.ex
@@ -5,7 +5,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 defmodule Qwen.Generation do
-  @moduledoc"""
+  @moduledoc """
   Qwen Generation
   """
   alias Qwen.Client
@@ -38,7 +38,6 @@ defmodule Qwen.Generation do
 
   ## 精简格式(默认)
   ```elixir
-  iex> import Qwen.Generation
   iex> prompt = ~l"model: qwen-turbo system: 你是一个学贯中西，通晓古今的文学家，给定一些历史上的文人，你能够根据这些人物的特征给出符合人物形象的对话。user: 你是唐代诗人李白，请做一首诗评价一下意大利作家卡尔维诺"
   [
     model: "qwen-turbo",
@@ -102,113 +101,4 @@ defmodule Qwen.Generation do
 
   # 解析其他异常返回
   defp parse_chat({:error, any_error}), do: {:error, any_error}
-
-  # sigil_l
-  @doc """
-  利用Elixir的Sigil，按照一定格式输入字符串，自动解析为请求需要的格式
-
-  ## 示例一: 一般用法
-  只需要指定`model`和`system`, `user`即可。
-
-  ```elixir
-  iex> import Qwen.Generation
-  iex> ~l"model: qwen-turbo system: 你是一个学贯中西，通晓古今的文学家，给定一些历史上的文人，你能够根据这些人物的特征给出符合人物形象的对话。user: 你是唐代诗人李白，请做一首诗评价一下意大利作家卡尔维诺"
-  [
-    model: "qwen-turbo",
-    input: %{
-      messages: [
-        %{
-          role: "system",
-          content: "你是一个学贯中西，通晓古今的文学家，给定一些历史上的文人，你能够根据这些人物的特征给出符合人物形象的对话。"
-        },
-        %{
-          role: "user",
-          content: "你是唐代诗人李白，请做一首诗评价一下意大利作家卡尔维诺"
-        }
-      ]
-    },
-    parameters: %{}
-  ]
-  ```
-
-  ## 示例二: 指定`parameters`
-  通过`parameters.key: value`指定`parameters`部分的参数`key`为`value`,
-  如加入`parameters.result_format: message`可以设置参数`parameters.result_format`为`message`
-
-  ```elixir
-  iex> import Qwen.Generation
-  iex> ~l"model: qwen-turbo parameters.result_format: message system: 你是一个学贯中西，通晓古今的文学家，给定一些历史上的文人，你能够根据这些人物的特征给出符合人物形象的对话。user: 你是唐代诗人李白，请做一首诗评价一下意大利作家卡尔维诺"
-  [
-    model: "qwen-turbo",
-    input: %{
-      messages: [
-        %{
-          role: "system",
-          content: "你是一个学贯中西，通晓古今的文学家，给定一些历史上的文人，你能够根据这些人物的特征给出符合人物形象的对话。"
-        },
-        %{
-          role: "user",
-          content: "你是唐代诗人李白，请做一首诗评价一下意大利作家卡尔维诺"
-        }
-      ]
-    },
-    parameters: %{result_format: "message"}
-  ]
-  ```
-  """
-  def sigil_l(lines, _opts) do
-    lines |> text_to_prompts()
-  end
-
-  defp text_to_prompts(text) when is_binary(text) do
-    model = extract_model(text) |> String.trim()
-    messages = extract_messages(text)
-    parameters = extract_parameters(text)
-    [model: model, input: %{messages: messages}, parameters: parameters]
-  end
-
-  defp extract_model(text) do
-    extract_value_after_keyword(text, "model:")
-  end
-
-  defp extract_messages(text) do
-    keywords = ["system:", "user:", "assistant:"]
-
-    Enum.reduce_while(keywords, [], fn keyword, acc ->
-      case extract_value_after_keyword(text, keyword) do
-        nil ->
-          {:cont, acc}
-
-        value ->
-          role = String.trim(keyword, ":")
-          acc = acc ++ [%{role: role, content: String.trim(value)}]
-          {:cont, acc}
-      end
-    end)
-  end
-
-  defp extract_parameters(text) do
-    keywords = ["result_format:"]
-
-    Enum.reduce_while(keywords, %{}, fn keyword, acc ->
-      case extract_value_after_keyword(text, "parameters." <> keyword) do
-        nil ->
-          {:cont, acc}
-
-        value ->
-          param = String.trim(keyword, ":")
-          acc = Map.put(acc, String.to_existing_atom(param), String.trim(value))
-          {:cont, acc}
-      end
-    end)
-  end
-
-  defp extract_value_after_keyword(text, keyword) do
-    pattern = ~r/#{keyword}\s*(.*?)(?=model:|system:|user:|assistant:|parameters|$)/s
-
-    case Regex.run(pattern, text) do
-      [_, value] -> value
-      _ -> nil
-    end
-  end
 end